Improved idempotence of the configure_dmc task
Add changed_when and failed_when directives to the shell module in
configure_dmc. The task will show a change when the rc is 0 and the
instance was added successfully. If the command returns a rc of 24 this
means the instance had already been added and the result will show ok.
All other rc will cause a failure.